requirements.txt
asgiref==3.7.2
dateutils==0.6.12
Django==4.2.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
